+++ src/Model/Resources/Wood.java
@@ -16,8 +16,8 @@ public class Wood extends RenewableResource {
     public static final String resourceName = "Wood";
 	public static final int gatheringTime = Constants.GATHER_WOOD_STATE_TIME;
 
-	private final float xPoss;
-	private final float yPoss;
+	private final double xPoss;
+	private final double yPoss;
 
 	private int yield;
 	private int updateCounter;
@@ -31,7 +31,7 @@ public class Wood extends RenewableResource {
 	 * @param maxResources The maximum amount of resources this object can contain at a time.
 	 * @param yield The amount of items gained from this resource at a time.
 	 */
-    public Wood(int initial, int maxResources, int yield, float xPoss, float yPoss){
+    public Wood(int initial, int maxResources, int yield, double xPoss, double yPoss){
         super(initial, maxResources);
 		this.yield = yield;
 		this.updateCounter = 0;
@@ -132,8 +132,8 @@ public class Wood extends RenewableResource {
 
 	@Override
 	public void spawn(World rhs) {
-		float xDiff = ((float) Math.random() * Constants.TREE_SPAWN_RADIUS) - Constants.TREE_SPAWN_RADIUS/2;
-		float yDiff = ((float) Math.random() * Constants.TREE_SPAWN_RADIUS) - Constants.TREE_SPAWN_RADIUS/2;
+		double xDiff = (Math.random() * Constants.TREE_SPAWN_RADIUS) - Constants.TREE_SPAWN_RADIUS/2;
+		double yDiff = (Math.random() * Constants.TREE_SPAWN_RADIUS) - Constants.TREE_SPAWN_RADIUS/2;
 		Wood wood = new Wood(Constants.WOOD_INITIAL_RESOURCES, Constants.WOOD_MAX_RESOURCES, Constants.WOOD_RESOURCE_GAIN, xPoss + xDiff, yPoss + yDiff);
 		rhs.addRenewableResourcePoint(wood, RenderObject.RENDER_OBJECT_ENUM.WOOD, xPoss + xDiff, yPoss + yDiff, Constants.TREE_COLLISION_RADIUS);
 		this.spawning = false;
